(almost) complete component layout refactor
