+++ custom_components/meross_lan/binary_sensor.py
@@ -2,7 +2,7 @@ import typing
 
 from homeassistant.components import binary_sensor
 
-from . import meross_entity as me
+from .helpers import entity as me
 
 if typing.TYPE_CHECKING:
     from homeassistant.config_entries import ConfigEntry
@@ -15,6 +15,6 @@ async def async_setup_entry(
     me.platform_setup_entry(hass, config_entry, async_add_devices, binary_sensor.DOMAIN)
 
 
-class MLBinarySensor(me.MerossBinaryEntity, binary_sensor.BinarySensorEntity):
+class MLBinarySensor(me.MLBinaryEntity, binary_sensor.BinarySensorEntity):
     PLATFORM = binary_sensor.DOMAIN
     DeviceClass = binary_sensor.BinarySensorDeviceClass
