+++ custom_components/meross_lan/devices/mts960.py
@@ -9,7 +9,7 @@ from ..number import MLEmulatedNumber
 from ..sensor import MLDiagnosticSensor
 
 if typing.TYPE_CHECKING:
-    from ..meross_device import MerossDevice
+    from ..helpers.device import Device
     from ..number import MtsTemperatureNumber
 
 
@@ -46,7 +46,7 @@ class Mts960PlugState(MLBinarySensor):
 class Mts960Climate(MtsClimate):
     """Climate entity for MTS960 devices"""
 
-    manager: "MerossDevice"
+    manager: "Device"
     ns = mn.Appliance_Control_Thermostat_ModeB
     device_scale = mc.MTS960_TEMP_SCALE
 
@@ -103,7 +103,7 @@ class Mts960Climate(MtsClimate):
 
     def __init__(
         self,
-        manager: "MerossDevice",
+        manager: "Device",
         channel: object,
         adjust_number_class: typing.Type["MtsTemperatureNumber"],
     ):
@@ -128,7 +128,7 @@ class Mts960Climate(MtsClimate):
             self, "timer_cycle_on_duration"
         )
 
-    # interface: MerossEntity
+    # interface: MLEntity
     async def async_shutdown(self):
         await super().async_shutdown()
         self.binary_sensor_plug_state: Mts960PlugState = None  # type: ignore
