+++ custom_components/meross_lan/diagnostics.py
@@ -1,33 +1,37 @@
 import typing
 
-from . import MerossApi, const as mlc
+from . import const as mlc
 from .helpers import ConfigEntryType
+from .helpers.component_api import ComponentApi
+from .helpers.meross_profile import MerossProfile, MerossProfileStore
 from .helpers.obfuscate import OBFUSCATE_DEVICE_ID_MAP, obfuscated_dict
-from .meross_profile import MerossCloudProfile, MerossCloudProfileStore
 
 if typing.TYPE_CHECKING:
+    from typing import Any, Mapping
+
     from homeassistant.config_entries import ConfigEntry
+    from homeassistant.core import HomeAssistant
 
 
 async def async_get_device_diagnostics(
-    hass, config_entry: "ConfigEntry", device
-) -> typing.Mapping[str, typing.Any]:
+    hass: "HomeAssistant", config_entry: "ConfigEntry", device
+) -> "Mapping[str, Any]":
     """Return diagnostics for a device entry."""
     return await async_get_config_entry_diagnostics(hass, config_entry)
 
 
 async def async_get_config_entry_diagnostics(
-    hass, config_entry: "ConfigEntry"
-) -> typing.Mapping[str, typing.Any]:
+    hass: "HomeAssistant", config_entry: "ConfigEntry"
+) -> "Mapping[str, Any]":
     """Return diagnostics for a config entry."""
-
+    api = ComponentApi.get(hass)
     data = config_entry.data
     obfuscate = data.get(mlc.CONF_OBFUSCATE, True)
     data = obfuscated_dict(data) if obfuscate else dict(data)
 
     match ConfigEntryType.get_type_and_id(config_entry.unique_id):
         case (ConfigEntryType.DEVICE, device_id):
-            if device := MerossApi.devices.get(device_id):
+            if device := api.devices.get(device_id):
                 data["device"] = {
                     "class": type(device).__name__,
                     "conf_protocol": device.conf_protocol,
@@ -37,9 +41,7 @@ async def async_get_config_entry_diagnostics(
                     "device_response_size_min": device.device_response_size_min,
                     "device_response_size_max": device.device_response_size_max,
                     "MQTT": {
-                        "cloud_profile": isinstance(
-                            device._profile, MerossCloudProfile
-                        ),
+                        "cloud_profile": isinstance(device._profile, MerossProfile),
                         "locally_active": bool(device.mqtt_locallyactive),
                         "mqtt_connection": bool(device._mqtt_connection),
                         "mqtt_connected": bool(device._mqtt_connected),
@@ -78,11 +80,11 @@ async def async_get_config_entry_diagnostics(
             return data
 
         case (ConfigEntryType.PROFILE, profile_id):
-            if profile := MerossApi.profiles.get(profile_id):
+            if profile := api.profiles.get(profile_id):
                 store_data = profile._data
             else:
                 try:
-                    store_data = await MerossCloudProfileStore(profile_id).async_load()
+                    store_data = await MerossProfileStore(hass, profile_id).async_load()
                 except Exception:
                     store_data = None
             if obfuscate:
@@ -91,10 +93,10 @@ async def async_get_config_entry_diagnostics(
                 # is not smart enough (but OBFUSCATE_DEVICE_ID_MAP is already
                 # filled with uuid(s) from the profile device_info(s) and
                 # the device_info(s) were already obfuscated in data)
-                store_data[MerossCloudProfile.KEY_DEVICE_INFO] = {
+                store_data[MerossProfile.KEY_DEVICE_INFO] = {
                     OBFUSCATE_DEVICE_ID_MAP[device_id]: device_info
                     for device_id, device_info in store_data[
-                        MerossCloudProfile.KEY_DEVICE_INFO
+                        MerossProfile.KEY_DEVICE_INFO
                     ].items()
                 }
             data["store"] = store_data
