+++ custom_components/meross_lan/helpers/meross_profile.py
@@ -2,699 +2,57 @@
 meross_lan module interface to access Meross Cloud services
 """
 
-import abc
 import asyncio
 from contextlib import asynccontextmanager
 from time import time
 import typing
 
-from homeassistant.config_entries import SOURCE_INTEGRATION_DISCOVERY
 from homeassistant.core import callback
 from homeassistant.helpers import storage
 from homeassistant.util import dt as dt_util
 
-from . import const as mlc, meross_entity as me
-from .const import (
-    CONF_CHECK_FIRMWARE_UPDATES,
-    CONF_DEVICE_ID,
-    CONF_KEY,
-    CONF_PASSWORD,
-    CONF_PAYLOAD,
-    DOMAIN,
-    DeviceConfigType,
-)
-from .devices.hub import HubMixin
-from .helpers import (
-    ConfigEntriesHelper,
-    Loggable,
+from . import (
     datetime_from_epoch,
     get_default_ssl_context,
     versiontuple,
 )
-from .helpers.manager import ApiProfile, CloudApiClient
-from .merossclient import (
+from .. import const as mlc
+from ..const import (
+    CONF_CHECK_FIRMWARE_UPDATES,
+    CONF_PASSWORD,
+    DOMAIN,
+)
+from ..merossclient import (
     MEROSSDEBUG,
     HostAddress,
-    MerossKeyError,
-    MerossRequest,
-    MerossResponse,
-    check_message_strict,
     const as mc,
     get_active_broker,
-    get_message_uuid,
-    get_replykey,
     namespaces as mn,
 )
-from .merossclient.cloudapi import APISTATUS_TOKEN_ERRORS, CloudApiError
-from .merossclient.mqttclient import (
-    MerossMQTTAppClient,
-    MerossMQTTRateLimitException,
-    generate_app_id,
-)
-from .repairs import IssueSeverity, create_issue, remove_issue
-from .sensor import MLDiagnosticSensor
+from ..merossclient.cloudapi import APISTATUS_TOKEN_ERRORS, CloudApiError
+from ..merossclient.mqttclient import MerossMQTTAppClient, generate_app_id
+from .manager import CloudApiClient
+from .mqtt_profile import ConnectionSensor, MQTTConnection, MQTTProfile
 
 if typing.TYPE_CHECKING:
+    from typing import Unpack
 
-    from homeassistant.components import mqtt as ha_mqtt
     from homeassistant.config_entries import ConfigEntry
-    from homeassistant.helpers.service_info.mqtt import MqttServiceInfo
-    import paho.mqtt.client as paho_mqtt
+    from homeassistant.core import HomeAssistant
 
-    from . import MerossApi
-    from .const import ProfileConfigType
-    from .meross_device import MerossDevice, MerossDeviceDescriptor
-    from .merossclient import (
-        MerossHeaderType,
-        MerossMessage,
-        MerossMessageType,
-        MerossPayloadType,
-    )
-    from .merossclient.cloudapi import (
+    from ..const import ProfileConfigType
+    from ..devices.hub import HubMixin
+    from ..merossclient import MerossMessage
+    from ..merossclient.cloudapi import (
         DeviceInfoType,
         LatestVersionType,
         SubDeviceInfoType,
     )
+    from .component_api import ComponentApi
+    from .device import Device, MerossDeviceDescriptor
 
     UuidType = str
     DeviceInfoDictType = dict[UuidType, "DeviceInfoType"]
-
-
-class ConnectionSensor(me.MEAlwaysAvailableMixin, MLDiagnosticSensor):
-    STATE_DISCONNECTED: typing.Final = "disconnected"
-    STATE_CONNECTED: typing.Final = "connected"
-    STATE_DROPPING: typing.Final = "dropping"
-
-    class AttrDictType(typing.TypedDict):
-        devices: dict[str, str]
-        received: int
-        published: int
-        dropped: int
-
-    ATTR_DEVICES: typing.Final = "devices"
-    ATTR_RECEIVED: typing.Final = "received"
-    ATTR_PUBLISHED: typing.Final = "published"
-    ATTR_DROPPED: typing.Final = "dropped"
-
-    manager: ApiProfile
-
-    # HA core entity attributes:
-    extra_state_attributes: AttrDictType
-    _unrecorded_attributes = frozenset(
-        {
-            ATTR_DEVICES,
-            ATTR_RECEIVED,
-            ATTR_PUBLISHED,
-            ATTR_DROPPED,
-            *MLDiagnosticSensor._unrecorded_attributes,
-        }
-    )
-    native_value: str
-    options: list[str] = [
-        STATE_DISCONNECTED,
-        STATE_CONNECTED,
-        STATE_DROPPING,
-    ]
-
-    __slots__ = ("connection",)
-
-    def __init__(self, connection: "MQTTConnection"):
-        self.connection = connection
-        self.extra_state_attributes = {
-            ConnectionSensor.ATTR_DEVICES: {
-                device.id: device.name for device in connection.mqttdevices.values()
-            },
-            ConnectionSensor.ATTR_RECEIVED: 0,
-            ConnectionSensor.ATTR_PUBLISHED: 0,
-            ConnectionSensor.ATTR_DROPPED: 0,
-        }
-        super().__init__(
-            connection.profile,
-            None,
-            connection.id,
-            native_value=(
-                self.STATE_CONNECTED
-                if connection.mqtt_is_connected
-                else self.STATE_DISCONNECTED
-            ),
-        )
-        connection.sensor_connection = self
-
-    # interface: Loggable
-    def configure_logger(self):
-        self.logtag = (
-            f"{self.__class__.__name__}({self.manager.loggable_broker(self.id)})"
-        )
-
-    # interface: MLDiagnosticSensor
-    async def async_shutdown(self):
-        await super().async_shutdown()
-        self.connection.sensor_connection = None
-        self.connection: MQTTConnection = None  # type: ignore
-
-    # interface: self
-    def update_devices(self):
-        # rebuild the attr (sub)dict else we were keeping a reference
-        # to the underlying hass.state and updates were missing
-        self.extra_state_attributes[ConnectionSensor.ATTR_DEVICES] = {
-            device.id: device.name for device in self.connection.mqttdevices.values()
-        }
-        self.flush_state()
-
-    def inc_counter(self, attr_name: str):
-        self.extra_state_attributes[attr_name] += 1
-        self.flush_state()
-
-    def inc_counter_with_state(self, attr_name: str, state: str):
-        self.extra_state_attributes[attr_name] += 1
-        self.native_value = state
-        self.flush_state()
-
-
-class _MQTTTransaction:
-    """Context for pending MQTT publish(es) waiting for responses.
-    This will allow to synchronize message request-response flow on MQTT
-    """
-
-    __slots__ = (
-        "mqtt_connection",
-        "device_id",
-        "namespace",
-        "messageid",
-        "method",
-        "request_time",
-        "response_future",
-    )
-
-    def __init__(
-        self,
-        mqtt_connection: "MQTTConnection",
-        device_id: str,
-        request: "MerossMessage",
-    ):
-        self.mqtt_connection = mqtt_connection
-        self.device_id = device_id
-        self.namespace = request.namespace
-        self.messageid = request.messageid
-        self.method = request.method
-        self.request_time = time()
-        self.response_future: "asyncio.Future[MerossResponse]" = (
-            asyncio.get_running_loop().create_future()
-        )
-        mqtt_connection._mqtt_transactions[request.messageid] = self
-
-    def cancel(self):
-        mqtt_connection = self.mqtt_connection
-        mqtt_connection.log(
-            mqtt_connection.DEBUG,
-            "Cancelling mqtt transaction on %s %s (uuid:%s messageId:%s)",
-            self.method,
-            self.namespace,
-            mqtt_connection.profile.loggable_device_id(self.device_id),
-            self.messageid,
-        )
-        self.response_future.cancel()
-        mqtt_connection._mqtt_transactions.pop(self.messageid, None)
-
-
-class MQTTConnection(Loggable):
-    """
-    Base abstract class representing a connection to an MQTT
-    broker. Historically, MQTT support was only through MerossApi
-    and the HA core MQTT broker. The introduction of Meross cloud
-    connection has 'generalized' the concept of the MQTT broker.
-    This interface is used by devices to actually send/receive
-    MQTT messages (in place of the legacy approach using MerossApi)
-    and represents a link to a broker (either through HA or a
-    merosss cloud mqtt)
-    """
-
-    _MQTT_DROP = "DROP"
-    _MQTT_PUBLISH = "PUBLISH"
-    _MQTT_RECV = "RECV"
-
-    DEFAULT_RESPONSE_TIMEOUT = 5
-
-    SESSION_HANDLERS: typing.Mapping[
-        str,
-        typing.Callable[
-            ["MQTTConnection", str, "MerossHeaderType", "MerossPayloadType"],
-            typing.Awaitable[bool],
-        ],
-    ] = {}
-
-    broker: "HostAddress"
-
-    __slots__ = (
-        "profile",
-        "broker",
-        "topic_response",
-        "mqttdevices",
-        "mqttdiscovering",
-        "namespace_handlers",
-        "sensor_connection",
-        "_mqtt_transactions",
-        "_mqtt_is_connected",
-    )
-
-    def __init__(
-        self,
-        profile: "MerossCloudProfile | MerossApi",
-        broker: "HostAddress",
-        topic_response: str,
-    ):
-        self.profile: typing.Final = profile
-        self.broker = broker
-        self.topic_response: typing.Final = topic_response
-        self.mqttdevices: typing.Final[dict[str, "MerossDevice"]] = {}
-        self.mqttdiscovering: typing.Final[set[str]] = set()
-        self.namespace_handlers = self.SESSION_HANDLERS
-        self.sensor_connection: "ConnectionSensor | None" = None
-        self._mqtt_transactions: typing.Final[dict[str, "_MQTTTransaction"]] = {}
-        self._mqtt_is_connected = False
-        super().__init__(
-            str(broker),
-            logger=profile,
-        )
-        profile.mqttconnections[self.id] = self
-        if profile.create_diagnostic_entities:
-            ConnectionSensor(self)
-
-    # interface: Loggable
-    def configure_logger(self):
-        self.logtag = (
-            f"{self.__class__.__name__}({self.profile.loggable_broker(self.broker)})"
-        )
-
-    # interface: self
-    async def async_shutdown(self):
-        for mqtt_transaction in list(self._mqtt_transactions.values()):
-            mqtt_transaction.cancel()
-        self.mqttdiscovering.clear()
-        for device in self.mqttdevices.values():
-            device.mqtt_detached()
-        self.mqttdevices.clear()
-        self.sensor_connection = None
-
-    async def async_create_diagnostic_entities(self):
-        if not self.sensor_connection:
-            ConnectionSensor(self)
-
-    async def entry_update_listener(self, profile: ApiProfile):
-        """Called by the ApiProfile to propagate config changes"""
-        self.configure_logger()
-        if self.sensor_connection:
-            self.sensor_connection.configure_logger()
-
-    @property
-    @abc.abstractmethod
-    def is_cloud_connection(self):
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def get_rl_safe_delay(self, uuid: str):
-        raise NotImplementedError()
-
-    @property
-    def mqtt_is_connected(self):
-        return self._mqtt_is_connected
-
-    def attach(self, device: "MerossDevice"):
-        assert device.id not in self.mqttdevices, (
-            "unexpected MQTTConnection.attach",
-            device.id,
-        )
-        device.mqtt_attached(self)
-        self.mqttdevices[device.id] = device
-        if sensor_connection := self.sensor_connection:
-            sensor_connection.update_devices()
-
-    def detach(self, device: "MerossDevice"):
-        device_id = device.id
-        assert device_id in self.mqttdevices, (
-            "unexpected MQTTConnection.detach",
-            device_id,
-        )
-        for mqtt_transaction in list(self._mqtt_transactions.values()):
-            if mqtt_transaction.device_id == device_id:
-                mqtt_transaction.cancel()
-        device.mqtt_detached()
-        self.mqttdevices.pop(device_id)
-        if sensor_connection := self.sensor_connection:
-            sensor_connection.update_devices()
-
-    @typing.final
-    def mqtt_publish(
-        self,
-        device_id: str,
-        request: "MerossMessage",
-    ):
-        return self.profile.async_create_task(
-            self.async_mqtt_publish(device_id, request), f".mqtt_publish({device_id})"
-        )
-
-    @typing.final
-    async def async_mqtt_publish(
-        self,
-        device_id: str,
-        request: "MerossMessage",
-    ) -> MerossResponse | None:
-        if request.method in mc.METHOD_ACK_MAP.keys():
-            transaction = _MQTTTransaction(self, device_id, request)
-        else:
-            transaction = None
-        try:
-            self.profile.trace_or_log(self, device_id, request, ApiProfile.TRACE_TX)
-            await self._async_mqtt_publish(device_id, request)
-            if transaction:
-                try:
-                    return await asyncio.wait_for(
-                        transaction.response_future, self.DEFAULT_RESPONSE_TIMEOUT
-                    )
-                except Exception as exception:
-                    self.log_exception(
-                        self.DEBUG,
-                        exception,
-                        "waiting for MQTT reply to %s %s (uuid:%s messageId:%s)",
-                        request.method,
-                        request.namespace,
-                        self.profile.loggable_device_id(device_id),
-                        request.messageid,
-                    )
-                finally:
-                    self._mqtt_transactions.pop(transaction.messageid, None)
-            return None
-
-        except MerossMQTTRateLimitException:
-            if sensor_connection := self.sensor_connection:
-                sensor_connection.inc_counter_with_state(
-                    ConnectionSensor.ATTR_DROPPED,
-                    ConnectionSensor.STATE_DROPPING,
-                )
-            self.log(
-                self.WARNING,
-                "MQTT publish rate-limit exceeded for device uuid:%s",
-                self.profile.loggable_device_id(device_id),
-            )
-
-        except Exception as exception:
-            self.log_exception(
-                self.WARNING,
-                exception,
-                "async_mqtt_publish %s %s (uuid:%s messageId:%s)",
-                request.method,
-                request.namespace,
-                self.profile.loggable_device_id(device_id),
-                request.messageid,
-                timeout=14400,
-            )
-
-        if transaction:
-            transaction.cancel()
-        return None
-
-    @typing.final
-    async def async_mqtt_message(
-        self,
-        mqtt_msg: "ha_mqtt.ReceiveMessage | paho_mqtt.MQTTMessage | MqttServiceInfo",
-    ):
-        with self.exception_warning("async_mqtt_message"):
-            if sensor_connection := self.sensor_connection:
-                sensor_connection.inc_counter(ConnectionSensor.ATTR_RECEIVED)
-            mqtt_payload = mqtt_msg.payload
-            message = MerossResponse(
-                mqtt_payload
-                if type(mqtt_payload) is str
-                else mqtt_payload.decode("utf-8")  # type: ignore
-            )
-            header = message[mc.KEY_HEADER]
-            device_id = get_message_uuid(header)
-            namespace = header[mc.KEY_NAMESPACE]
-            messageid = header[mc.KEY_MESSAGEID]
-            payload = message[mc.KEY_PAYLOAD]
-
-            profile = self.profile
-            profile.trace_or_log(self, device_id, message, ApiProfile.TRACE_RX)
-
-            if messageid in self._mqtt_transactions:
-                mqtt_transaction = self._mqtt_transactions[messageid]
-                if mqtt_transaction.namespace == namespace:
-                    self._mqtt_transactions.pop(messageid, None)
-                    mqtt_transaction.response_future.set_result(message)
-            else:
-                # special session management: cloud connections would
-                # behave differently than the local MQTT. Their behavior
-                # will definitevely be set in the dynamic/custom message handlers
-                # implemented in the derived MQTTConnections
-                if namespace in self.namespace_handlers:
-                    if await self.namespace_handlers[namespace](
-                        self, device_id, header, payload
-                    ):
-                        # session management has already taken care of everything
-                        return
-
-            try:
-                self.mqttdevices[device_id].mqtt_receive(message)
-                return
-            except KeyError:
-                # device is not binded to this MQTTConnection
-                if device := ApiProfile.devices.get(device_id):
-                    # check among current loaded devices if they could be re-binded
-                    if device.conf_protocol is mlc.CONF_PROTOCOL_HTTP:
-                        self.log(
-                            self.DEBUG,
-                            "Dropping MQTT received message for device uuid:%s since it is configured for HTTP only",
-                            profile.loggable_device_id(device_id),
-                        )
-                        return
-                    if device._profile == profile:
-                        self.attach(device)
-                    else:
-                        if (device.key != profile.key) or (
-                            device.descriptor.userId != profile.id
-                        ):
-                            # this is not really expected and deserves a warning but is expected
-                            # when you (re)bind a device and it still is connected to the old broker
-                            # until reboot
-                            self.log(
-                                self.WARNING,
-                                "Received MQTT message for device uuid:%s which cannot be registered for MQTT handling on this profile",
-                                profile.loggable_device_id(device_id),
-                                timeout=14400,
-                            )
-                            return
-                        profile.link(device)
-                        # profile.link will attach to the mqtt broker known to the device cfg..
-                        # we'll ensure that (in case device cfg is stale) we're correctly binded here
-                        if device._mqtt_connection != self:
-                            self.attach(device)
-
-                    device.mqtt_receive(message)
-                    return
-
-            # the device is not configured: proceed to discovery in case
-            if device_id in self.mqttdiscovering:
-                return
-
-            # lookout for any disabled/ignored entry
-            config_entries_helper = ConfigEntriesHelper(self.hass)
-            if (
-                (profile is self.api)
-                and (not config_entries_helper.get_config_entry(DOMAIN))
-                and (not config_entries_helper.get_config_flow(DOMAIN))
-            ):
-                # not really needed but we would like to always have the
-                # MQTT hub entry in case so if the user removed that..retrigger
-                await self.hass.config_entries.flow.async_init(
-                    DOMAIN,
-                    context={"source": "hub"},
-                    data=None,
-                )
-
-            if config_entry := (
-                config_entries_helper.get_config_entry(device_id)
-                or config_entries_helper.get_config_entry(device_id[-12:].lower())
-            ):
-                # entry already present...skip discovery
-                self.log(
-                    self.INFO,
-                    "Ignoring MQTT discovery for already configured uuid:%s (ConfigEntry is %s)",
-                    profile.loggable_device_id(device_id),
-                    (
-                        "disabled"
-                        if config_entry.disabled_by
-                        else "ignored" if config_entry.source == "ignore" else "unknown"
-                    ),
-                    timeout=28800,  # type: ignore
-                )
-                return
-
-            # also skip discovered integrations waiting in HA queue
-            if config_entries_helper.get_config_flow(device_id):
-                self.log(
-                    self.DEBUG,
-                    "Ignoring MQTT discovery for uuid:%s (ConfigFlow is in progress)",
-                    profile.loggable_device_id(device_id),
-                    timeout=14400,  # type: ignore
-                )
-                return
-
-            key = profile.key
-            if get_replykey(header, key) is not key:
-                self.log(
-                    self.WARNING,
-                    "Discovery key error for uuid:%s",
-                    profile.loggable_device_id(device_id),
-                    timeout=300,
-                )
-                if key is not None:
-                    return
-
-            self.profile.async_create_task(
-                self.async_try_discovery(device_id),
-                f".async_try_discovery({device_id})",
-            )
-
-    async def async_identify_device(self, device_id: str, key: str) -> DeviceConfigType:
-        """
-        Sends an ns_all and ns_ability GET requests encapsulated in an ns_multiple
-        to speed up things. Raises exception in case of error
-        """
-        try:
-            response = check_message_strict(
-                await self.async_mqtt_publish(
-                    device_id,
-                    MerossRequest(
-                        key,
-                        *mn.Appliance_System_Ability.request_get,
-                        self.topic_response,
-                    ),
-                )
-            )
-            ability = response[mc.KEY_PAYLOAD][mc.KEY_ABILITY]
-        except MerossKeyError as error:
-            raise error
-        except Exception as exception:
-            raise Exception("Unable to identify abilities") from exception
-
-        try:
-            response = check_message_strict(
-                await self.async_mqtt_publish(
-                    device_id,
-                    MerossRequest(
-                        key,
-                        *mn.Appliance_System_All.request_get,
-                        self.topic_response,
-                    ),
-                )
-            )
-            all = response[mc.KEY_PAYLOAD][mc.KEY_ALL]
-        except MerossKeyError as error:
-            raise error
-        except Exception as exception:
-            raise Exception("Unable to identify device (all)") from exception
-        return {
-            CONF_DEVICE_ID: device_id,
-            CONF_PAYLOAD: {
-                mc.KEY_ALL: all,
-                mc.KEY_ABILITY: ability,
-            },
-            CONF_KEY: key,
-        }
-
-    async def async_try_discovery(self, device_id: str):
-        """
-        Tries device identification and starts a flow if succeded returning
-        the FlowResult. Returns None if anything fails for whatever reason.
-        """
-
-        self.mqttdiscovering.add(device_id)
-        try:
-            result = await self.hass.config_entries.flow.async_init(
-                DOMAIN,
-                context={"source": SOURCE_INTEGRATION_DISCOVERY},
-                data=await self.async_identify_device(device_id, self.profile.key),
-            )
-        except Exception as e:
-            result = None
-            self.log_exception(
-                self.WARNING,
-                e,
-                "async_try_discovery (uuid:%s)",
-                self.profile.loggable_device_id(device_id),
-                timeout=14400,
-            )
-        self.mqttdiscovering.remove(device_id)
-        return result
-
-    def _mqtt_transactions_clean(self):
-        if self._mqtt_transactions:
-            # check and cleanup stale transactions
-            epoch = time()
-            for transaction in list(self._mqtt_transactions.values()):
-                if (epoch - transaction.request_time) > 15:
-                    transaction.cancel()
-
-    @abc.abstractmethod
-    async def _async_mqtt_publish(
-        self,
-        device_id: str,
-        request: "MerossMessage",
-    ):
-        """
-        Actually sends the message to the transport. On return gives
-        (status_code, timeout) with the expected timeout-to-reply depending
-        on the queuing system in place (MerossMQTTConnection/paho client).
-        Should raise an exception when the message could not be sent
-        """
-        raise NotImplementedError()
-
-    @callback
-    def _mqtt_connected(self):
-        """called when the underlying mqtt.Client connects to the broker"""
-        for device in self.mqttdevices.values():
-            device.mqtt_connected()
-        self._mqtt_is_connected = True
-        if sensor_connection := self.sensor_connection:
-            sensor_connection.update_native_value(ConnectionSensor.STATE_CONNECTED)
-
-    @callback
-    def _mqtt_disconnected(self):
-        """called when the underlying mqtt.Client disconnects from the broker"""
-        for device in self.mqttdevices.values():
-            device.mqtt_disconnected()
-        self._mqtt_is_connected = False
-        if sensor_connection := self.sensor_connection:
-            sensor_connection.update_native_value(ConnectionSensor.STATE_DISCONNECTED)
-
-    @callback
-    def _mqtt_published(self):
-        """called when the underlying mqtt.Client successfully publishes a message"""
-        if sensor_connection := self.sensor_connection:
-            sensor_connection.inc_counter(ConnectionSensor.ATTR_PUBLISHED)
-
-    async def _handle_Appliance_System_Online(
-        self, device_id: str, header: "MerossHeaderType", payload: "MerossPayloadType"
-    ):
-        """
-        This is likely sent by the session management layer on the Meross brokers
-        to notify the app of the device connection state. We then intercept
-        this message which is not intended for the device though and act accordingly
-        here at our 'session management state'. At any rate, this will be set to be
-        handled in every MQTTConnection (cloud, local) so we process even messages
-        originated from the device itself
-        """
-        if header[mc.KEY_METHOD] == mc.METHOD_PUSH:
-            status = payload[mc.KEY_ONLINE].get(mc.KEY_STATUS)
-            if status == mc.STATUS_ONLINE:
-                # the device is now online on this connection: tell the pipe to continue processing
-                # This will in turn (eventually) link the device to the current profile/connection
-                # (if not already) and online it since it will receive a 'fresh' MQTT
-                return False
-        # any other condition will instruct the message pipe
-        # to abort processing since the device is not online or we don't
-        # understand this message
-        return True
 
 
 class MerossMQTTConnection(MQTTConnection, MerossMQTTAppClient):
@@ -716,15 +74,16 @@ class MerossMQTTConnection(MQTTConnection, MerossMQTTAppClient):
         "_unsub_random_disconnect",
     )
 
-    def __init__(self, profile: "MerossCloudProfile", broker: "HostAddress"):
+    def __init__(self, profile: "MerossProfile", broker: "HostAddress"):
         MerossMQTTAppClient.__init__(
             self,
             profile.key,
             profile.userid,
             app_id=profile.app_id,
-            loop=self.hass.loop,
+            loop=profile.hass.loop,
             sslcontext=get_default_ssl_context(),
         )
+        self.is_cloud_connection = True
         MQTTConnection.__init__(self, profile, broker, self.topic_command)
         if profile.isEnabledFor(profile.VERBOSE):
             self.enable_logger(self)  # type: ignore (Loggable is duck-compatible with Logger)
@@ -759,16 +118,12 @@ class MerossMQTTConnection(MQTTConnection, MerossMQTTAppClient):
         await MerossMQTTAppClient.async_shutdown(self)
         await MQTTConnection.async_shutdown(self)
 
-    async def entry_update_listener(self, profile: "MerossCloudProfile"):
+    async def entry_update_listener(self, profile: "MerossProfile"):
         await MQTTConnection.entry_update_listener(self, profile)
         if profile.isEnabledFor(profile.VERBOSE):
             self.enable_logger(self)  # type: ignore (Loggable is duck-compatible with Logger)
         else:
             self.disable_logger()
-
-    @property
-    def is_cloud_connection(self):
-        return True
 
     def get_rl_safe_delay(self, uuid: str):
         return MerossMQTTAppClient.get_rl_safe_delay(self, uuid)
@@ -778,7 +133,7 @@ class MerossMQTTConnection(MQTTConnection, MerossMQTTAppClient):
         device_id: str,
         request: "MerossMessage",
     ):
-        return await self.hass.async_add_executor_job(
+        return await self.profile.hass.async_add_executor_job(
             self.rl_publish, device_id, request
         )
 
@@ -804,7 +159,7 @@ MerossMQTTConnection.SESSION_HANDLERS = {
 }
 
 
-class MerossCloudProfileStoreType(typing.TypedDict):
+class MerossProfileStoreType(typing.TypedDict):
     appId: str
     # TODO credentials: typing.NotRequired[MerossCloudCredentials]
     deviceInfo: "DeviceInfoDictType"
@@ -815,22 +170,27 @@ class MerossCloudProfileStoreType(typing.TypedDict):
     tokenRequestTime: float
 
 
-class MerossCloudProfileStore(storage.Store[MerossCloudProfileStoreType]):
+class MerossProfileStore(storage.Store[MerossProfileStoreType]):
     VERSION = 1
 
-    def __init__(self, profile_id: str):
+    def __init__(self, hass: "HomeAssistant", profile_id: str):
         super().__init__(
-            ApiProfile.hass,
-            MerossCloudProfileStore.VERSION,
+            hass,
+            MerossProfileStore.VERSION,
             f"{DOMAIN}.profile.{profile_id}",
         )
 
 
-class MerossCloudProfile(ApiProfile):
+class MerossProfile(MQTTProfile):
     """
     Represents and manages a cloud account profile used to retrieve keys
     and/or to manage cloud mqtt connection(s)
     """
+
+    if typing.TYPE_CHECKING:
+        config: ProfileConfigType
+        _data: MerossProfileStoreType
+        _unsub_polling_query_device_info: asyncio.TimerHandle | None
 
     KEY_APP_ID: typing.Final = "appId"
     KEY_DEVICE_INFO: typing.Final = "deviceInfo"
@@ -840,9 +200,6 @@ class MerossCloudProfile(ApiProfile):
     KEY_LATEST_VERSION_TIME: typing.Final = "latestVersionTime"
     KEY_TOKEN_REQUEST_TIME: typing.Final = "tokenRequestTime"
 
-    config: "ProfileConfigType"
-    _data: "MerossCloudProfileStoreType"
-
     __slots__ = (
         "apiclient",
         "_data",
@@ -851,8 +208,12 @@ class MerossCloudProfile(ApiProfile):
         "_device_info_time",
     )
 
-    def __init__(self, profile_id: str, config_entry: "ConfigEntry"):
-        ApiProfile.__init__(self, profile_id, config_entry)
+    def __init__(
+        self, profile_id: str, api: "ComponentApi", config_entry: "ConfigEntry"
+    ):
+        MQTTProfile.__init__(
+            self, profile_id, api=api, hass=api.hass, config_entry=config_entry
+        )
         # state of the art for credentials is that they're mixed in
         # into the config_entry.data but this is prone to issues and confusing
         # so we 'might' decide to move them to a dict valued key in configentry.data
@@ -862,8 +223,8 @@ class MerossCloudProfile(ApiProfile):
         # to change the version(s) in storage/config. At the moment I'm still very confused
         # and opting to keep the credentials where they are embedded in ConfigEntry
         self.apiclient = CloudApiClient(self, self.config)
-        self._store = MerossCloudProfileStore(profile_id)
-        self._unsub_polling_query_device_info: asyncio.TimerHandle | None = None
+        self._store = MerossProfileStore(self.hass, profile_id)
+        self._unsub_polling_query_device_info = None
 
     async def async_init(self):
         """
@@ -899,7 +260,7 @@ class MerossCloudProfile(ApiProfile):
                 await self.async_token_refresh()
         else:
             self._device_info_time = 0.0
-            self._data: MerossCloudProfileStoreType = {
+            self._data: MerossProfileStoreType = {
                 self.KEY_APP_ID: generate_app_id(),
                 mc.KEY_TOKEN: self.config.get(mc.KEY_TOKEN),
                 self.KEY_DEVICE_INFO: {},
@@ -932,12 +293,12 @@ class MerossCloudProfile(ApiProfile):
             self._unsub_polling_query_device_info.cancel()
             self._unsub_polling_query_device_info = None
         await super().async_shutdown()
-        ApiProfile.profiles[self.id] = None
+        self.api.profiles[self.id] = None
 
     # interface: ConfigEntryManager
     async def entry_update_listener(self, hass, config_entry: "ConfigEntry"):
         config: ProfileConfigType = config_entry.data  # type: ignore
-        remove_issue(mlc.ISSUE_CLOUD_TOKEN_EXPIRED, self.id)
+        self.remove_issue(mlc.ISSUE_CLOUD_TOKEN_EXPIRED)
         curr_credentials = self.apiclient.credentials
         if not curr_credentials or (
             curr_credentials[mc.KEY_TOKEN] != config[mc.KEY_TOKEN]
@@ -951,7 +312,7 @@ class MerossCloudProfile(ApiProfile):
                 await self._store.async_save(self._data)
 
         if self.config.get(mc.KEY_MQTTDOMAIN) != config.get(mc.KEY_MQTTDOMAIN):
-            self.schedule_entry_reload()
+            self.schedule_reload()
         else:
             await super().entry_update_listener(hass, config_entry)
             # the 'async_check_query_devices' will only occur if we didn't refresh
@@ -965,7 +326,7 @@ class MerossCloudProfile(ApiProfile):
         return f"profile_{self.loggable_profile_id(self.id)}"
 
     # interface: ApiProfile
-    def attach_mqtt(self, device: "MerossDevice"):
+    def attach_mqtt(self, device: "Device"):
         descr = device.descriptor
         try:
             if device.online:
@@ -1025,7 +386,7 @@ class MerossCloudProfile(ApiProfile):
     def userid(self):
         return self.config[mc.KEY_USERID_]
 
-    def link(self, device: "MerossDevice"):
+    def link(self, device: "Device"):
         """
         Device linking to a cloud profile sets the environment for
         the device MQTT attachment/connection. This process uses a lot
@@ -1134,7 +495,7 @@ class MerossCloudProfile(ApiProfile):
                     await self.apiclient.async_device_latestversion()
                 )
                 self._schedule_save_store()
-                for device in ApiProfile.active_devices():
+                for device in self.api.active_devices():
                     if latest_version := self.get_latest_version(device.descriptor):
                         device.update_latest_version(latest_version)
 
@@ -1223,24 +584,22 @@ class MerossCloudProfile(ApiProfile):
             # since it should still be pending...
             data[mc.KEY_TOKEN] = credentials[mc.KEY_TOKEN]
             self.log(self.INFO, "Meross api token was automatically refreshed")
-            helper = ConfigEntriesHelper(self.hass)
-            profile_entry = helper.get_config_entry(f"profile.{self.id}")
+            profile_entry = self.api.get_config_entry(f"profile.{self.id}")
             if profile_entry:
                 # weird enough if this isnt true...
                 profile_config = dict(profile_entry.data)
                 profile_config.update(credentials)
                 # watchout: this will in turn call self.entry_update_listener
-                helper.config_entries.async_update_entry(
+                self.hass.config_entries.async_update_entry(
                     profile_entry,
                     data=profile_config,
                 )
             return credentials
         except Exception as exception:
             self.log_exception(self.WARNING, exception, "Meross api token auto-refresh")
-            create_issue(
+            self.create_issue(
                 mlc.ISSUE_CLOUD_TOKEN_EXPIRED,
-                self.id,
-                severity=IssueSeverity.WARNING,
+                severity=self.IssueSeverity.WARNING,
                 translation_placeholders={"email": config.get(mc.KEY_EMAIL)},
             )
             return None
@@ -1296,6 +655,7 @@ class MerossCloudProfile(ApiProfile):
     async def _process_device_info_new(
         self, device_info_list_new: list["DeviceInfoType"]
     ):
+        api_devices = self.api.devices
         device_info_dict = self._data[self.KEY_DEVICE_INFO]
         device_info_removed = {device_id for device_id in device_info_dict.keys()}
         device_info_unknown: list["DeviceInfoType"] = []
@@ -1316,15 +676,14 @@ class MerossCloudProfile(ApiProfile):
                     sub_device_info_dict = None
                 device_info_dict[device_id] = device_info
 
-                if device_id not in ApiProfile.devices:
+                try:
+                    device = api_devices[device_id]
+                except KeyError:
                     device_info_unknown.append(device_info)
                     continue
-
-                if (device := ApiProfile.devices[device_id]) is None:
-                    # config_entry for device is not loaded
+                if not device:  # device loaded
                     continue
-
-                if isinstance(device, HubMixin):
+                if device.get_type() is mlc.DeviceType.HUB:
                     if sub_device_info_dict is None:
                         sub_device_info_dict = {}
                     device_info[self.KEY_SUBDEVICE_INFO] = sub_device_info_dict
@@ -1333,7 +692,9 @@ class MerossCloudProfile(ApiProfile):
                     )
                     if sub_device_info_list_new is not None:
                         await self._process_subdevice_info_new(
-                            device, sub_device_info_dict, sub_device_info_list_new
+                            typing.cast("HubMixin", device),
+                            sub_device_info_dict,
+                            sub_device_info_list_new,
                         )
                 device.update_device_info(device_info)
 
@@ -1394,7 +755,6 @@ class MerossCloudProfile(ApiProfile):
             )
             return
 
-        config_entries_helper = ConfigEntriesHelper(self.hass)
         for device_info in device_info_unknown:
             with self.exception_warning("_process_device_info_unknown"):
                 device_id = device_info[mc.KEY_UUID]
@@ -1403,7 +763,7 @@ class MerossCloudProfile(ApiProfile):
                     "Trying/Initiating discovery for (new) uuid:%s",
                     self.loggable_device_id(device_id),
                 )
-                if config_entries_helper.get_config_flow(device_id):
+                if self.api.get_config_flow(device_id):
                     continue  # device configuration already progressing
                 # cloud conf has a new device
                 if domain := device_info.get(mc.KEY_DOMAIN):
