+++ custom_components/meross_lan/number.py
@@ -2,20 +2,17 @@ import typing
 
 from homeassistant.components import number
 
-from . import meross_entity as me
-from .helpers import reverse_lookup
+from .helpers import entity as me, reverse_lookup
 from .merossclient import const as mc
 
 if typing.TYPE_CHECKING:
+    from typing import Unpack
+
     from homeassistant.config_entries import ConfigEntry
     from homeassistant.core import HomeAssistant
 
     from .climate import MtsClimate
-    from .meross_device import MerossDeviceBase
-
-    # optional arguments for MLConfigNumber init
-    class MLConfigNumberArgs(me.MerossNumericEntityArgs):
-        pass
+    from .helpers.device import BaseDevice
 
 
 async def async_setup_entry(
@@ -24,7 +21,7 @@ async def async_setup_entry(
     me.platform_setup_entry(hass, config_entry, async_add_devices, number.DOMAIN)
 
 
-class MLNumber(me.MerossNumericEntity, number.NumberEntity):
+class MLNumber(me.MLNumericEntity, number.NumberEntity):
     """
     Base (abstract) ancestor for ML number entities. This has 2 specializations:
     - MLConfigNumber: for configuration parameters backed by a device namespace value.
@@ -40,15 +37,15 @@ class MLNumber(me.MerossNumericEntity, number.NumberEntity):
 
     DEVICECLASS_TO_UNIT_MAP = {
         None: None,
-        DEVICE_CLASS_DURATION: me.MerossEntity.hac.UnitOfTime.SECONDS,
-        DeviceClass.HUMIDITY: me.MerossEntity.hac.PERCENTAGE,
-        DeviceClass.TEMPERATURE: me.MerossEntity.hac.UnitOfTemperature.CELSIUS,
+        DEVICE_CLASS_DURATION: me.MLEntity.hac.UnitOfTime.SECONDS,
+        DeviceClass.HUMIDITY: me.MLEntity.hac.PERCENTAGE,
+        DeviceClass.TEMPERATURE: me.MLEntity.hac.UnitOfTemperature.CELSIUS,
     }
 
-    manager: "MerossDeviceBase"
+    manager: "BaseDevice"
 
     # HA core entity attributes:
-    entity_category = me.EntityCategory.CONFIG
+    entity_category = me.MLNumericEntity.EntityCategory.CONFIG
     mode: number.NumberMode = number.NumberMode.BOX
     native_max_value: float
     native_min_value: float
@@ -72,11 +69,11 @@ class MLConfigNumber(me.MEListChannelMixin, MLNumber):
 
     def __init__(
         self,
-        manager: "MerossDeviceBase",
+        manager: "BaseDevice",
         channel: object | None,
         entitykey: str | None = None,
         device_class: MLNumber.DeviceClass | str | None = None,
-        **kwargs: "typing.Unpack[MLConfigNumberArgs]",
+        **kwargs: "Unpack[MLConfigNumber.Args]",
     ):
         self._async_request_debounce_unsub = None
         super().__init__(
@@ -158,7 +155,7 @@ class MtsTemperatureNumber(MLConfigNumber):
         self,
         climate: "MtsClimate",
         entitykey: str,
-        **kwargs: "typing.Unpack[MLConfigNumberArgs]",
+        **kwargs: "Unpack[MLConfigNumber.Args]",
     ):
         self.climate = climate
         kwargs["device_scale"] = climate.device_scale
