+++ custom_components/meross_lan/update.py
@@ -2,26 +2,26 @@ import typing
 
 from homeassistant.components import update
 
-from . import meross_entity as me
+from .helpers import entity as me
 from .merossclient import const as mc
 from .merossclient.cloudapi import LatestVersionType
 
 if typing.TYPE_CHECKING:
 
-    from .meross_device import MerossDevice
+    from .helpers.device import Device
 
 
 async def async_setup_entry(hass, config_entry, async_add_devices):
     me.platform_setup_entry(hass, config_entry, async_add_devices, update.DOMAIN)
 
 
-class MLUpdate(me.MEAlwaysAvailableMixin, me.MerossEntity, update.UpdateEntity):
+class MLUpdate(me.MEAlwaysAvailableMixin, me.MLEntity, update.UpdateEntity):
     PLATFORM = update.DOMAIN
     DeviceClass = update.UpdateDeviceClass
-    manager: "MerossDevice"
+    manager: "Device"
 
     # HA core entity attributes:
-    entity_category = me.EntityCategory.DIAGNOSTIC
+    entity_category = me.MLEntity.EntityCategory.DIAGNOSTIC
     installed_version: str | None
     latest_version: str | None
     release_summary: str | None
@@ -32,7 +32,7 @@ class MLUpdate(me.MEAlwaysAvailableMixin, me.MerossEntity, update.UpdateEntity):
         "release_summary",
     )
 
-    def __init__(self, manager: "MerossDevice", latest_version: LatestVersionType):
+    def __init__(self, manager: "Device", latest_version: LatestVersionType):
         self.installed_version = manager.descriptor.firmwareVersion
         self.latest_version = latest_version.get(mc.KEY_VERSION)
         self.release_summary = latest_version.get(mc.KEY_DESCRIPTION)
